moving tasks tests to celery app and pep8 ed tasks.py
